src/marcador-reports/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyodbc==5.1.0
reportlab==4.2.2
python-jose[cryptography]==3.3.0
Pillow==10.4.0